Chore(Question): Refactor
